Switch Sliders and ProgressBars to floating point values

+++ src/Eto.WinForms/Forms/Controls/ProgressBarHandler.cs
@@ -25,19 +25,19 @@ namespace Eto.WinForms.Forms.Controls
 			}
 		}
 
-		public int MaxValue {
+		public double MaxValue {
 			get { return Control.Maximum; }
-			set { Control.Maximum = value; }
+			set { Control.Maximum = (int)value; }
 		}
 
-		public int MinValue {
+		public double MinValue {
 			get { return Control.Minimum; }
-			set { Control.Minimum = value; }
+			set { Control.Minimum = (int)value; }
 		}
 
-		public int Value {
+		public double Value {
 			get { return Control.Value; }
-			set { Control.Value = value; }
+			set { Control.Value = (int)value; }
 		}
 	}
 }
